requirement.txt
faiss-cpu
torch
sentence-transformers
langchain
langchain-groq
python-dotenv
groq